requirements.txt
discord.py[voice]==2.3.2
Flask==2.3.2